+++ src/rules/rule-no-duplicate-properties/__tests__/index.js
@@ -14,7 +14,8 @@ testRule(undefined, tr => {
   tr.ok("a { color: pink; { &:hover { color: orange; } } }", "spec nested")
   tr.ok("a { color: pink; @media { color: orange; } }", "nested")
   tr.ok("a { color: pink; @media { color: orange; &::before { color: black; } } }", "double nested")
-
+  tr.ok("@fontface { src: url(font.eof); src: url(font.woff) }")
+  
   tr.notOk("a { color: pink; color: orange }", messages.rejected("color"))
   tr.notOk("a { color: pink; background: orange; color: orange }", messages.rejected("color"))
   tr.notOk("a { color: pink; background: orange; background: pink; }", messages.rejected("background"))